+++ projects/amdsmi/py-interface/amdsmi_wrapper.py
@@ -1298,7 +1298,8 @@ struct_amdsmi_asic_info_t._fields_ = [
     ('subsystem_id', ctypes.c_uint32),
     ('PADDING_1', ctypes.c_ubyte * 4),
     ('flags', ctypes.c_uint64),
-    ('reserved', ctypes.c_uint32 * 18),
+    ('physical_acc_id', ctypes.c_uint32),
+    ('reserved', ctypes.c_uint32 * 17),
 ]
 
 amdsmi_asic_info_t = struct_amdsmi_asic_info_t
@@ -2702,6 +2703,31 @@ struct_amdsmi_npm_info_t._fields_ = [
 ]
 
 amdsmi_npm_info_t = struct_amdsmi_npm_info_t
+
+# values for enumeration 'amdsmi_compute_tray_type_t'
+amdsmi_compute_tray_type_t__enumvalues = {
+    0: 'AMDSMI_COMPUTE_TRAY_TYPE_UNKNOWN',
+    1: 'AMDSMI_COMPUTE_TRAY_TYPE_HELIOS_P',
+    2: 'AMDSMI_COMPUTE_TRAY_TYPE_HELIOS_R',
+    3: 'AMDSMI_COMPUTE_TRAY_TYPE_TITAN',
+}
+AMDSMI_COMPUTE_TRAY_TYPE_UNKNOWN = 0
+AMDSMI_COMPUTE_TRAY_TYPE_HELIOS_P = 1
+AMDSMI_COMPUTE_TRAY_TYPE_HELIOS_R = 2
+AMDSMI_COMPUTE_TRAY_TYPE_TITAN = 3
+amdsmi_compute_tray_type_t = ctypes.c_uint32 # enum
+class struct_amdsmi_tray_info_t(Structure):
+    pass
+
+struct_amdsmi_tray_info_t._pack_ = 1 # source:False
+struct_amdsmi_tray_info_t._layout_ = 'ms'
+struct_amdsmi_tray_info_t._fields_ = [
+    ('max_acc_per_tray', ctypes.c_uint32),
+    ('tray_type', amdsmi_compute_tray_type_t),
+    ('reserved', ctypes.c_uint32 * 14),
+]
+
+amdsmi_tray_info_t = struct_amdsmi_tray_info_t
 
 # values for enumeration 'amdsmi_ptl_data_format_t'
 amdsmi_ptl_data_format_t__enumvalues = {
@@ -4287,6 +4313,12 @@ try:
 except AttributeError:
     pass
 try:
+    amdsmi_get_tray_info = _libraries['libamd_smi.so'].amdsmi_get_tray_info
+    amdsmi_get_tray_info.restype = amdsmi_status_t
+    amdsmi_get_tray_info.argtypes = [amdsmi_node_handle, ctypes.POINTER(struct_amdsmi_tray_info_t)]
+except AttributeError:
+    pass
+try:
     amdsmi_get_fw_info = _libraries['libamd_smi.so'].amdsmi_get_fw_info
     amdsmi_get_fw_info.restype = amdsmi_status_t
     amdsmi_get_fw_info.argtypes = [amdsmi_processor_handle, ctypes.POINTER(struct_amdsmi_fw_info_t)]
@@ -4946,7 +4978,11 @@ __all__ = \
     'AMDSMI_COMPUTE_PARTITION_MEM_ALLOC_CAPPING',
     'AMDSMI_COMPUTE_PARTITION_MEM_ALLOC_INVALID',
     'AMDSMI_COMPUTE_PARTITION_QPX', 'AMDSMI_COMPUTE_PARTITION_SPX',
-    'AMDSMI_COMPUTE_PARTITION_TPX', 'AMDSMI_CONTAINER_DOCKER',
+    'AMDSMI_COMPUTE_PARTITION_TPX',
+    'AMDSMI_COMPUTE_TRAY_TYPE_HELIOS_P',
+    'AMDSMI_COMPUTE_TRAY_TYPE_HELIOS_R',
+    'AMDSMI_COMPUTE_TRAY_TYPE_TITAN',
+    'AMDSMI_COMPUTE_TRAY_TYPE_UNKNOWN', 'AMDSMI_CONTAINER_DOCKER',
     'AMDSMI_CONTAINER_LXC', 'AMDSMI_CPER_NOTIFY_TYPE_BOOT',
     'AMDSMI_CPER_NOTIFY_TYPE_CMC', 'AMDSMI_CPER_NOTIFY_TYPE_CPE',
     'AMDSMI_CPER_NOTIFY_TYPE_CXL_COMPONENT',
@@ -5260,13 +5296,13 @@ __all__ = \
     'amdsmi_card_form_factor_t', 'amdsmi_clean_gpu_local_data',
     'amdsmi_clk_info_t', 'amdsmi_clk_limit_type_t',
     'amdsmi_clk_type_t', 'amdsmi_compute_partition_mem_alloc_mode_t',
-    'amdsmi_compute_partition_type_t', 'amdsmi_container_types_t',
-    'amdsmi_counter_command_t', 'amdsmi_counter_value_t',
-    'amdsmi_cper_guid_t', 'amdsmi_cper_hdr_t',
-    'amdsmi_cper_notify_type_t', 'amdsmi_cper_sev_t',
-    'amdsmi_cper_timestamp_t', 'amdsmi_cper_valid_bits_t',
-    'amdsmi_cpu_apb_disable', 'amdsmi_cpu_apb_enable',
-    'amdsmi_cpu_info_t', 'amdsmi_cpu_util_t',
+    'amdsmi_compute_partition_type_t', 'amdsmi_compute_tray_type_t',
+    'amdsmi_container_types_t', 'amdsmi_counter_command_t',
+    'amdsmi_counter_value_t', 'amdsmi_cper_guid_t',
+    'amdsmi_cper_hdr_t', 'amdsmi_cper_notify_type_t',
+    'amdsmi_cper_sev_t', 'amdsmi_cper_timestamp_t',
+    'amdsmi_cper_valid_bits_t', 'amdsmi_cpu_apb_disable',
+    'amdsmi_cpu_apb_enable', 'amdsmi_cpu_info_t', 'amdsmi_cpu_util_t',
     'amdsmi_cpusocket_handle', 'amdsmi_ddr_bw_metrics_t',
     'amdsmi_dev_perf_level_t', 'amdsmi_dimm_power_t',
     'amdsmi_dimm_thermal_t', 'amdsmi_dpm_level_t',
@@ -5405,11 +5441,11 @@ __all__ = \
     'amdsmi_get_soc_pstate', 'amdsmi_get_socket_handles',
     'amdsmi_get_socket_info', 'amdsmi_get_supported_power_cap',
     'amdsmi_get_temp_metric', 'amdsmi_get_threads_per_core',
-    'amdsmi_get_ttm_info', 'amdsmi_get_utilization_count',
-    'amdsmi_get_vcn_busy_percent', 'amdsmi_get_violation_status',
-    'amdsmi_get_xgmi_info', 'amdsmi_get_xgmi_plpd',
-    'amdsmi_gpu_block_t', 'amdsmi_gpu_cache_info_t',
-    'amdsmi_gpu_control_counter',
+    'amdsmi_get_tray_info', 'amdsmi_get_ttm_info',
+    'amdsmi_get_utilization_count', 'amdsmi_get_vcn_busy_percent',
+    'amdsmi_get_violation_status', 'amdsmi_get_xgmi_info',
+    'amdsmi_get_xgmi_plpd', 'amdsmi_gpu_block_t',
+    'amdsmi_gpu_cache_info_t', 'amdsmi_gpu_control_counter',
     'amdsmi_gpu_counter_group_supported', 'amdsmi_gpu_create_counter',
     'amdsmi_gpu_destroy_counter', 'amdsmi_gpu_metrics_t',
     'amdsmi_gpu_ras_policy_info_t', 'amdsmi_gpu_ras_policy_v4_0_t',
@@ -5483,8 +5519,9 @@ __all__ = \
     'amdsmi_temperature_type_t', 'amdsmi_topo_get_link_type',
     'amdsmi_topo_get_link_weight', 'amdsmi_topo_get_numa_node_number',
     'amdsmi_topo_get_p2p_status', 'amdsmi_topology_nearest_t',
-    'amdsmi_ttm_info_t', 'amdsmi_uma_carveout_info_t',
-    'amdsmi_uma_carveout_option_t', 'amdsmi_utilization_counter_t',
+    'amdsmi_tray_info_t', 'amdsmi_ttm_info_t',
+    'amdsmi_uma_carveout_info_t', 'amdsmi_uma_carveout_option_t',
+    'amdsmi_utilization_counter_t',
     'amdsmi_utilization_counter_type_t', 'amdsmi_vbios_info_t',
     'amdsmi_version_t', 'amdsmi_violation_status_t',
     'amdsmi_virtualization_mode_t', 'amdsmi_voltage_metric_t',
@@ -5548,8 +5585,8 @@ __all__ = \
     'struct_amdsmi_retired_page_record_t',
     'struct_amdsmi_smu_fw_version_t', 'struct_amdsmi_sock_info_t',
     'struct_amdsmi_temp_range_refresh_rate_t',
-    'struct_amdsmi_topology_nearest_t', 'struct_amdsmi_ttm_info_t',
-    'struct_amdsmi_uma_carveout_info_t',
+    'struct_amdsmi_topology_nearest_t', 'struct_amdsmi_tray_info_t',
+    'struct_amdsmi_ttm_info_t', 'struct_amdsmi_uma_carveout_info_t',
     'struct_amdsmi_uma_carveout_option_t',
     'struct_amdsmi_utilization_counter_t',
     'struct_amdsmi_vbios_info_t', 'struct_amdsmi_version_t',