+++ projects/amdsmi/tests/amd_smi_test/main.cc
@@ -63,6 +63,7 @@
 #include "functional/gpu/xgmi/xgmi_read_write.h"
 #include "functional/ifoe/fabric/fabric_read.h"
 #include "functional/ifoe/identity/ifoe_info_read.h"
+#include "functional/ifoe/tray/tray_info_read.h"
 #include "functional/system/cross_process_serialization.h"
 #include "functional/system/hw_topology_read.h"
 #include "functional/system/init_shutdown_refcount.h"
@@ -345,6 +346,11 @@ TEST(IfoeFunctionalReadOnly, TestFabricRead) {
 
 TEST(IfoeFunctionalReadOnly, TestIfoeInfoRead) {
   TestIfoeInfoRead tst;
+  RunGenericTest(&tst);
+}
+
+TEST(IfoeFunctionalReadOnly, TestTrayInfoRead) {
+  TestTrayInfoRead tst;
   RunGenericTest(&tst);
 }